Refactor: Board HOC 추가 및 적용
- CSS 박살!
