+++ src/Router.js
@@ -1,7 +1,6 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import { PageLayout } from '@common/PageLayout';
-import BoardList from './components/pages/articles/ArticleListPage';
 import ErrorPage from './components/pages/Error';
 import Login from './components/pages/login/Login';
 import Main from './components/pages/main/Main';
@@ -9,7 +8,8 @@ import Mypage from './components/pages/Mypage';
 import Auth from './components/pages/Auth';
 import Donation from './components/pages/Donation';
 import Writing from './components/pages/writing';
-import ArticleDetail from './components/pages/articles/ArticleDetailPage';
+import ArticleDetailPage from './components/pages/articles/ArticleDetailPage';
+import ArticleListPage from './components/pages/articles/ArticleListPage';
 
 const Router = () => {
   return (
@@ -17,12 +17,12 @@ const Router = () => {
       <Routes>
         <Route element={<PageLayout />}>
           <Route exact path="/" element={<Main />} />
-          <Route path="category" element={<BoardList />} />
-          <Route path="category/:id" element={<BoardList />} />
+          <Route path="category" element={<ArticleListPage />} />
+          <Route path="category/:id" element={<ArticleListPage />} />
           <Route path="mypage" element={<Mypage />} />
           <Route path="mypage/:articleType" element={<Mypage />} />
           <Route path="writing" element={<Writing />} />
-          <Route path="article/:id" element={<ArticleDetail />} />
+          <Route path="article/:id" element={<ArticleDetailPage />} />
           <Route path="/donation" element={<Donation />} />
         </Route>
