+++ src/components/pages/articles/common/styles.jsx
@@ -1,0 +1,8 @@
+/** @jsxImportSource @emotion/react */
+
+import { css } from '@emotion/react';
+
+export const block = css`
+  margin: 0 0.8rem;
+  width: 100%;
+`;
